Move WidgetPod and BaseState into widget_pod.rs
This is the first bit of a larger refactor that will move code
out of lib.rs.
