+++ druid/src/lib.rs
@@ -33,17 +33,17 @@ mod menu;
 mod mouse;
 pub mod theme;
 pub mod widget;
+mod widget_pod;
 mod win_handler;
 mod window;
 
+use log;
 use std::collections::VecDeque;
 use std::ops::{Deref, DerefMut};
 use std::time::Instant;
 
-use log::{error, warn};
-
-use kurbo::{Affine, Point, Rect, Shape, Size, Vec2};
-use piet::{Piet, RenderContext};
+use kurbo::{Rect, Size};
+use piet::Piet;
 
 // these are the types from shell that we expose; others we only use internally.
 pub use shell::{
@@ -62,77 +62,12 @@ pub use lens::{Lens, LensExt, LensWrap};
 pub use localization::LocalizedString;
 pub use menu::{sys as platform_menus, ContextMenu, MenuDesc, MenuItem};
 pub use mouse::MouseEvent;
+pub use widget_pod::{BaseState, WidgetPod};
 pub use win_handler::DruidHandler;
 pub use window::{Window, WindowId};
 
-/// A container for one widget in the hierarchy.
-///
-/// Generally, container widgets don't contain other widgets directly,
-/// but rather contain a `WidgetPod`, which has additional state needed
-/// for layout and for the widget to participate in event flow.
-///
-/// This struct also contains the previous data for a widget, which is
-/// essential for the [`update`] method, both to decide when the update
-/// needs to propagate, and to provide the previous data so that a
-/// widget can process a diff between the old value and the new.
-///
-/// [`update`]: trait.Widget.html#tymethod.update
-pub struct WidgetPod<T: Data, W: Widget<T>> {
-    state: BaseState,
-    old_data: Option<T>,
-    env: Option<Env>,
-    inner: W,
-}
-
 /// Convenience type for dynamic boxed widget.
 pub type BoxedWidget<T> = WidgetPod<T, Box<dyn Widget<T>>>;
-
-/// Generic state for all widgets in the hierarchy.
-///
-/// This struct contains the widget's layout rect, flags
-/// indicating when the widget is active or focused, and other
-/// state necessary for the widget to participate in event
-/// flow.
-///
-/// It is provided to [`paint`] calls as a non-mutable reference,
-/// largely so a widget can know its size, also because active
-/// and focus state can affect the widget's appearance. Other than
-/// that, widgets will generally not interact with it directly,
-/// but it is an important part of the [`WidgetPod`] struct.
-///
-/// [`paint`]: trait.Widget.html#tymethod.paint
-/// [`WidgetPod`]: struct.WidgetPod.html
-#[derive(Default)]
-pub struct BaseState {
-    layout_rect: Rect,
-
-    // TODO: consider using bitflags for the booleans.
-
-    // This should become an invalidation rect.
-    needs_inval: bool,
-
-    is_hot: bool,
-
-    is_active: bool,
-
-    /// Any descendant is active.
-    has_active: bool,
-
-    /// Any descendant has requested an animation frame.
-    request_anim: bool,
-
-    /// Any descendant has requested a timer.
-    ///
-    /// Note: we don't have any way of clearing this request, as it's
-    /// likely not worth the complexity.
-    request_timer: bool,
-
-    /// This widget or a descendant has focus.
-    has_focus: bool,
-
-    /// This widget or a descendant has requested focus.
-    request_focus: bool,
-}
 
 /// The trait implemented by all widgets.
 ///
@@ -407,357 +342,6 @@ pub struct BoxConstraints {
     max: Size,
 }
 
-impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
-    /// Create a new widget pod.
-    ///
-    /// In a widget hierarchy, each widget is wrapped in a `WidgetPod`
-    /// so it can participate in layout and event flow. The process of
-    /// adding a child widget to a container should call this method.
-    pub fn new(inner: W) -> WidgetPod<T, W> {
-        WidgetPod {
-            state: Default::default(),
-            old_data: None,
-            env: None,
-            inner,
-        }
-    }
-
-    /// Return a reference to the inner widget.
-    pub fn widget(&self) -> &W {
-        &self.inner
-    }
-
-    /// Return a mutable reference to the inner widget.
-    pub fn widget_mut(&mut self) -> &mut W {
-        &mut self.inner
-    }
-
-    /// Set layout rectangle.
-    ///
-    /// Intended to be called on child widget in container's `layout`
-    /// implementation.
-    pub fn set_layout_rect(&mut self, layout_rect: Rect) {
-        self.state.layout_rect = layout_rect;
-    }
-
-    /// Get the layout rectangle.
-    ///
-    /// This will be same value as set by `set_layout_rect`.
-    pub fn get_layout_rect(&self) -> Rect {
-        self.state.layout_rect
-    }
-
-    /// Paint a child widget.
-    ///
-    /// Generally called by container widgets as part of their [`paint`]
-    /// method.
-    ///
-    /// Note that this method does not apply the offset of the layout rect.
-    /// If that is desired, use [`paint_with_offset`] instead.
-    ///
-    /// [`layout`]: trait.Widget.html#method.layout
-    /// [`paint`]: trait.Widget.html#method.paint
-    /// [`paint_with_offset`]: #method.paint_with_offset
-    pub fn paint(&mut self, paint_ctx: &mut PaintCtx, data: &T, env: &Env) {
-        self.inner.paint(paint_ctx, &self.state, data, &env);
-    }
-
-    /// Paint the widget, translating it by the origin of its layout rectangle.
-    ///
-    /// This will recursively paint widgets, stopping if a widget's layout
-    /// rect is outside of the currently visible region.
-    // Discussion: should this be `paint` and the other `paint_raw`?
-    pub fn paint_with_offset(&mut self, paint_ctx: &mut PaintCtx, data: &T, env: &Env) {
-        self.paint_with_offset_impl(paint_ctx, data, env, false)
-    }
-
-    /// Paint the widget, even if its layout rect is outside of the currently
-    /// visible region.
-    pub fn paint_with_offset_always(&mut self, paint_ctx: &mut PaintCtx, data: &T, env: &Env) {
-        self.paint_with_offset_impl(paint_ctx, data, env, true)
-    }
-
-    /// Shared implementation that can skip drawing non-visible content.
-    fn paint_with_offset_impl(
-        &mut self,
-        paint_ctx: &mut PaintCtx,
-        data: &T,
-        env: &Env,
-        paint_if_not_visible: bool,
-    ) {
-        if !paint_if_not_visible && !paint_ctx.region().intersects(self.state.layout_rect) {
-            return;
-        }
-
-        if let Err(e) = paint_ctx.save() {
-            error!("saving render context failed: {:?}", e);
-            return;
-        }
-
-        let layout_origin = self.state.layout_rect.origin().to_vec2();
-        paint_ctx.transform(Affine::translate(layout_origin));
-
-        let visible = paint_ctx.region().to_rect() - layout_origin;
-
-        paint_ctx.with_child_ctx(visible, |ctx| {
-            self.inner.paint(ctx, &self.state, data, &env)
-        });
-
-        if let Err(e) = paint_ctx.restore() {
-            error!("restoring render context failed: {:?}", e);
-        }
-    }
-
-    /// Compute layout of a widget.
-    ///
-    /// Generally called by container widgets as part of their [`layout`]
-    /// method.
-    ///
-    /// [`layout`]: trait.Widget.html#method.layout
-    pub fn layout(
-        &mut self,
-        layout_ctx: &mut LayoutCtx,
-        bc: &BoxConstraints,
-        data: &T,
-        env: &Env,
-    ) -> Size {
-        self.inner.layout(layout_ctx, bc, data, &env)
-    }
-
-    /// Propagate an event.
-    ///
-    /// Generally the [`event`] method of a container widget will call this
-    /// method on all its children. Here is where a great deal of the event
-    /// flow logic resides, particularly whether to continue propagating
-    /// the event.
-    ///
-    /// [`event`]: trait.Widget.html#method.event
-    pub fn event(&mut self, ctx: &mut EventCtx, event: &Event, data: &mut T, env: &Env) {
-        // TODO: factor as much logic as possible into monomorphic functions.
-        if ctx.is_handled || !event.recurse() {
-            // This function is called by containers to propagate an event from
-            // containers to children. Non-recurse events will be invoked directly
-            // from other points in the library.
-            return;
-        }
-        let had_active = self.state.has_active;
-        let mut child_ctx = EventCtx {
-            win_ctx: ctx.win_ctx,
-            cursor: ctx.cursor,
-            command_queue: ctx.command_queue,
-            window: &ctx.window,
-            window_id: ctx.window_id,
-            base_state: &mut self.state,
-            had_active,
-            is_handled: false,
-            is_root: false,
-        };
-        let rect = child_ctx.base_state.layout_rect;
-        // Note: could also represent this as `Option<Event>`.
-        let mut recurse = true;
-        let mut hot_changed = None;
-        let child_event = match event {
-            Event::LifeCycle(event) => Event::LifeCycle(*event),
-            Event::Size(size) => {
-                recurse = ctx.is_root;
-                Event::Size(*size)
-            }
-            Event::MouseDown(mouse_event) => {
-                let had_hot = child_ctx.base_state.is_hot;
-                let now_hot = rect.winding(mouse_event.pos) != 0;
-                if (!had_hot) && now_hot {
-                    child_ctx.base_state.is_hot = true;
-                    hot_changed = Some(true);
-                }
-                recurse = had_active || !ctx.had_active && now_hot;
-                let mut mouse_event = mouse_event.clone();
-                mouse_event.pos -= rect.origin().to_vec2();
-                Event::MouseDown(mouse_event)
-            }
-            Event::MouseUp(mouse_event) => {
-                recurse = had_active || !ctx.had_active && rect.winding(mouse_event.pos) != 0;
-                let mut mouse_event = mouse_event.clone();
-                mouse_event.pos -= rect.origin().to_vec2();
-                Event::MouseUp(mouse_event)
-            }
-            Event::MouseMoved(mouse_event) => {
-                let had_hot = child_ctx.base_state.is_hot;
-                child_ctx.base_state.is_hot = rect.winding(mouse_event.pos) != 0;
-                if had_hot != child_ctx.base_state.is_hot {
-                    hot_changed = Some(child_ctx.base_state.is_hot);
-                }
-                recurse = had_active || had_hot || child_ctx.base_state.is_hot;
-                let mut mouse_event = mouse_event.clone();
-                mouse_event.pos -= rect.origin().to_vec2();
-                Event::MouseMoved(mouse_event)
-            }
-            Event::KeyDown(e) => {
-                recurse = child_ctx.base_state.has_focus;
-                Event::KeyDown(*e)
-            }
-            Event::KeyUp(e) => {
-                recurse = child_ctx.base_state.has_focus;
-                Event::KeyUp(*e)
-            }
-            Event::Paste(e) => {
-                recurse = child_ctx.base_state.has_focus;
-                Event::Paste(e.clone())
-            }
-            Event::Wheel(wheel_event) => {
-                recurse = had_active || child_ctx.base_state.is_hot;
-                Event::Wheel(wheel_event.clone())
-            }
-            Event::Zoom(zoom) => {
-                recurse = had_active || child_ctx.base_state.is_hot;
-                Event::Zoom(*zoom)
-            }
-            Event::HotChanged(is_hot) => Event::HotChanged(*is_hot),
-            Event::FocusChanged(_is_focused) => {
-                let had_focus = child_ctx.base_state.has_focus;
-                let focus = child_ctx.base_state.request_focus;
-                child_ctx.base_state.request_focus = false;
-                child_ctx.base_state.has_focus = focus;
-                recurse = focus || had_focus;
-                Event::FocusChanged(focus)
-            }
-            Event::AnimFrame(interval) => {
-                recurse = child_ctx.base_state.request_anim;
-                child_ctx.base_state.request_anim = false;
-                Event::AnimFrame(*interval)
-            }
-            Event::Timer(id) => {
-                recurse = child_ctx.base_state.request_timer;
-                Event::Timer(*id)
-            }
-            Event::Command(cmd) => Event::Command(cmd.clone()),
-        };
-        child_ctx.base_state.needs_inval = false;
-        if let Some(is_hot) = hot_changed {
-            let hot_changed_event = Event::HotChanged(is_hot);
-            self.inner
-                .event(&mut child_ctx, &hot_changed_event, data, &env);
-        }
-        if recurse {
-            child_ctx.base_state.has_active = false;
-            self.inner.event(&mut child_ctx, &child_event, data, &env);
-            child_ctx.base_state.has_active |= child_ctx.base_state.is_active;
-        };
-        ctx.base_state.needs_inval |= child_ctx.base_state.needs_inval;
-        ctx.base_state.request_anim |= child_ctx.base_state.request_anim;
-        ctx.base_state.request_timer |= child_ctx.base_state.request_timer;
-        ctx.base_state.is_hot |= child_ctx.base_state.is_hot;
-        ctx.base_state.has_active |= child_ctx.base_state.has_active;
-        ctx.base_state.request_focus |= child_ctx.base_state.request_focus;
-        ctx.is_handled |= child_ctx.is_handled;
-    }
-
-    /// Propagate a data update.
-    ///
-    /// Generally called by container widgets as part of their [`update`]
-    /// method.
-    ///
-    /// [`update`]: trait.Widget.html#method.update
-    pub fn update(&mut self, ctx: &mut UpdateCtx, data: &T, env: &Env) {
-        let data_same = if let Some(ref old_data) = self.old_data {
-            old_data.same(data)
-        } else {
-            false
-        };
-        let env_same = if let Some(ref old_env) = self.env {
-            old_env.same(env)
-        } else {
-            false
-        };
-
-        if data_same && env_same {
-            return;
-        }
-        self.inner.update(ctx, self.old_data.as_ref(), data, env);
-        self.old_data = Some(data.clone());
-        self.env = Some(env.clone());
-    }
-}
-
-impl<T: Data, W: Widget<T> + 'static> WidgetPod<T, W> {
-    /// Box the contained widget.
-    ///
-    /// Convert a `WidgetPod` containing a widget of a specific concrete type
-    /// into a dynamically boxed widget.
-    pub fn boxed(self) -> BoxedWidget<T> {
-        WidgetPod {
-            state: self.state,
-            old_data: self.old_data,
-            env: self.env,
-            inner: Box::new(self.inner),
-        }
-    }
-}
-
-impl BaseState {
-    /// The "hot" (aka hover) status of a widget.
-    ///
-    /// A widget is "hot" when the mouse is hovered over it. Widgets will
-    /// often change their appearance as a visual indication that they
-    /// will respond to mouse interaction.
-    ///
-    /// The hot status is computed from the widget's layout rect. In a
-    /// container hierarchy, all widgets with layout rects containing the
-    /// mouse position have hot status.
-    ///
-    /// Discussion: there is currently some confusion about whether a
-    /// widget can be considered hot when some other widget is active (for
-    /// example, when clicking to one widget and dragging to the next).
-    /// The documentation should clearly state the resolution.
-    pub fn is_hot(&self) -> bool {
-        self.is_hot
-    }
-
-    /// The active status of a widget.
-    ///
-    /// Active status generally corresponds to a mouse button down. Widgets
-    /// with behavior similar to a button will call [`set_active`] on mouse
-    /// down and then up.
-    ///
-    /// When a widget is active, it gets mouse events even when the mouse
-    /// is dragged away.
-    ///
-    /// [`set_active`]: struct.EventCtx.html#method.set_active
-    pub fn is_active(&self) -> bool {
-        self.is_active
-    }
-
-    /// The focus status of a widget.
-    ///
-    /// Focus means that the widget receives keyboard events.
-    ///
-    /// A widget can request focus using the [`request_focus`] method.
-    /// This will generally result in a separate event propagation of
-    /// a `FocusChanged` method, including sending `false` to the previous
-    /// widget that held focus.
-    ///
-    /// Only one leaf widget at a time has focus. However, in a container
-    /// hierarchy, all ancestors of that leaf widget are also invoked with
-    /// `FocusChanged(true)`.
-    ///
-    /// Discussion question: is "is_focused" a better name?
-    ///
-    /// [`request_focus`]: struct.EventCtx.html#method.request_focus
-    pub fn has_focus(&self) -> bool {
-        self.has_focus
-    }
-
-    /// The layout size.
-    ///
-    /// This is the layout size as ultimately determined by the parent
-    /// container. Generally it will be the same as the size returned by
-    /// the child widget's [`layout`] method.
-    ///
-    /// [`layout`]: trait.Widget.html#tymethod.layout
-    pub fn size(&self) -> Size {
-        self.layout_rect.size()
-    }
-}
-
 impl BoxConstraints {
     /// Create a new box constraints object.
     ///
@@ -820,8 +404,8 @@ impl BoxConstraints {
             && 0.0 <= self.min.height
             && self.min.height <= self.max.height)
         {
-            warn!("Bad BoxConstraints passed to {}:", name);
-            warn!("{:?}", self);
+            log::warn!("Bad BoxConstraints passed to {}:", name);
+            log::warn!("{:?}", self);
         }
     }
 
@@ -889,7 +473,7 @@ impl<'a, 'b> EventCtx<'a, 'b> {
     ///
     /// See [`BaseState::is_hot`](struct.BaseState.html#method.is_hot).
     pub fn is_hot(&self) -> bool {
-        self.base_state.is_hot
+        self.base_state.is_hot()
     }
 
     /// Query the "active" state of the widget.
@@ -925,7 +509,7 @@ impl<'a, 'b> EventCtx<'a, 'b> {
     ///
     /// See [`BaseState::has_focus`](struct.BaseState.html#method.has_focus).
     pub fn has_focus(&self) -> bool {
-        self.base_state.has_focus
+        self.base_state.has_focus()
     }
 
     /// Request keyboard focus.
